initilize v2 branch

+++ src/main/java/com/project/shopapp/repositories/RoleRepository.java
@@ -1,7 +1,7 @@
 package com.project.shopapp.repositories;
 
-import com.project.shopapp.models.Role;
+import com.project.shopapp.models.Entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 }
